core: stdcm: move postprocessing into a new class
